# api-gateway/docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: |-
    Gateway for Mallow algorithmic trading system.

    All protected routes require a Bearer JWT token obtained from the Identity service (`/api/v1/auth/login`).
  title: Mallow API Gateway
  version: "1.0"
paths:
...
  /api/v1/strategies:
    get:
      description: Returns all ~80 named strategy IDs that herald supports.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - BearerAuth: []
      summary: List available strategies
      tags:
      - Herald
  /api/v1/strategist/ui/:
    get:
      description: Serves the Google ADK development UI for the trading assistant
        agent.
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Strategist ADK Web UI
      tags:
      - Strategist
...
  /api/v1/symbols:
    get:
      description: Returns all symbols currently ingested by herald.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - BearerAuth: []
      summary: List live symbols
      tags:
      - Herald
  /api/v1/watch:
    get:
      description: Returns admin warm-set configs (indicator bootstrapping on startup).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: List watch configs
      tags:
      - Herald Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Watch config
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a watch config
      tags:
      - Herald Admin
  /api/v1/watch/{id}:
    delete:
      parameters:
      - description: Watch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a watch config
      tags:
      - Herald Admin
    get:
      parameters:
      - description: Watch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a watch config
      tags:
      - Herald Admin
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: JWT token — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"